feat: consolidate config, add buffer config
